Breaking out PDCMetadata classes into thier own files
Also removing rubocop violations
