+++ app/models/pdc_metadata/title.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+# Class for storing a title in our local representation
+module PDCMetadata
+  # value:      "100 años de soledad"
+  # title_type: "TranslatedTitle"
+  class Title
+    attr_accessor :title, :title_type
+    def initialize(title:, title_type: nil)
+      @title = title
+      @title_type = title_type
+    end
+
+    def main?
+      @title_type.blank?
+    end
+  end
+end
